+++ src/lib.rs
@@ -6,3 +6,4 @@ pub mod io;
 pub mod settings;
 pub mod storage;
 pub mod units;
+pub mod builders;
