+++ src/main.rs
@@ -1,6 +1,7 @@
 extern crate chrono;
 extern crate serde;
 
+mod builders;
 mod errors;
 mod helpers;
 mod io;
@@ -11,9 +12,10 @@ mod units;
 use crate::helpers::Deferrable;
 use io::IOKind;
 use std::sync::Arc;
+use crate::builders::pubsub_builder;
 
 use crate::errors::Result;
-use crate::io::{Direction, Publisher, PublisherInstance, ThresholdNotifier};
+use crate::io::{Comparison, Publisher, PublisherInstance, ThresholdNotifier};
 use crate::settings::Settings;
 use crate::storage::{Persistent, PollGroup};
 
@@ -43,21 +45,9 @@ fn main() -> Result<()> {
 
     // build subscribers/commands
     println!("\nBuilding subscribers ...");
-    for (id, &ref input) in poller.inputs.iter() {
-        let binding = PublisherInstance::default();
-        let publisher = binding.deferred();
-
-        input.try_lock().unwrap().add_publisher(publisher.clone()).unwrap();
-
-        let notifier = ThresholdNotifier::new(
-            format!("subscriber for {}", id),
-            1.0,
-            publisher.clone(),
-            Direction::Above,
-        );
-        let deferred = notifier.deferred();
-        let mut binding = publisher.try_lock().unwrap();
-        binding.subscribe(deferred);
+    for (id, input) in poller.inputs.iter() {
+        let name = format!("Subscriber for Input:{}", id);
+        pubsub_builder(input.clone(), name, 1.0, Comparison::GT)
     }
     println!("... Finished building\n");
