+++ src/storage/grouping.rs
@@ -1,10 +1,11 @@
 use chrono::{DateTime, Duration, Utc};
 use std::sync::Arc;
+use crate::builders::input_log_builder;
 use crate::helpers::Deferred;
 use crate::errors::Result;
-use crate::helpers::{check_results, input_log_builder};
+use crate::helpers::check_results;
 use crate::io::{InputContainer, PublisherInstance};
-use crate::io::{IOKind, IdType};
+use crate::io::{IdType, IOKind};
 use crate::settings::Settings;
 use crate::storage::{LogContainer, MappedCollection, Persistent};
